backend/requirements.txt
fastapi
uvicorn
pydantic

# AI Services
anthropic
openai

# File Parsing
PyPDF2
python-docx
python-multipart

# Environment & Utilities
python-dotenv

# Web Search & Scraping
wikipedia-api
googlesearch-python
requests
beautifulsoup4
lxml
trafilatura
httpx

# RAG & Vector Database
chromadb
sentence-transformers
langchain
langchain-community

# Caching
redis
hiredis

# Rate Limiting
slowapi

# XML Parsing for arXiv
xmltodict
